tests for flow and job names

check, assertion, and test for flow and job name: both are characters of length 1 and both exist in flows.

Usage

check_flow_job_name(flow_name, job_name)

assert_flow_job_name(
  flow_name,
  job_name,
  .var.name = checkmate::vname(flow_name),
  add = NULL
)

test_flow_job_name(flow_name, job_name)

Arguments

flow_name

character flow name

job_name

character job name

.var.name

character name of the checked object to print in assertions. Defaults to the heuristic implemented in vname.

add

AssertCollection collection to store assertion messages. See AssertCollection.

Details

check_flow_job_name

Functions

check_flow_job_name(): returns TRUE if check is true, character error message if check is false

assert_flow_job_name(): returns TRUE if check is true, throws error if check is false

test_flow_job_name(): returns TRUE if check is true, FALSE otherwise